dont write lock file when index already exists

+++ scripts/variant_effect_predictor/variant_effect_predictor.pl
@@ -1622,29 +1622,40 @@ sub setup_fasta() {
     -RANK => 1,
   );
   
-  debug("Checking/creating FASTA index") unless defined($config->{quiet});
-  
   # check lock file
   my $lock_file = $config->{fasta};
   $lock_file .= -d $config->{fasta} ? '/.vep.lock' : '.vep.lock';
   
   # lock file exists, indexing failed
   if(-e $lock_file) {
-    for(qw(.fai .index .vep.lock)) {
+    for(qw(.fai .index /directory.index /directory.fai .vep.lock)) {
       unlink($config->{fasta}.$_) if -e $config->{fasta}.$_;
     }
   }
   
+  my $index_exists = 0;
+
+  for my $fn(map {$config->{fasta}.$_} qw(.fai .index /directory.index /directory.fai)) {
+    if(-e $fn) {
+      $index_exists = 1;
+      last;
+    }
+  }
+
   # create lock file
-  open LOCK, ">$lock_file" or die("ERROR: Could not write to FASTA lock file $lock_file\n");
-  print LOCK "1\n";
-  close LOCK;
+  unless($index_exists) {
+    debug("Creating FASTA index") unless defined($config->{quiet});
+
+    open LOCK, ">$lock_file" or die("ERROR: Could not write to FASTA lock file $lock_file\n");
+    print LOCK "1\n";
+    close LOCK;
+  }
   
   # run indexing
   $config->{fasta_db} = Bio::DB::Fasta->new($config->{fasta});
   
   # remove lock file
-  unlink($lock_file);
+  unlink($lock_file) unless $index_exists;
 }
 
 sub setup_custom {
